+++ src/main/java/com/github/lessjava/types/inference/impl/HMTypeBase.java
@@ -4,7 +4,7 @@ import com.github.lessjava.types.inference.HMType;
 
 public class HMTypeBase extends HMType {
     public static final HMTypeBase INT = new HMTypeBase(BaseDataType.INT);
-    public static final HMTypeBase REAL = new HMTypeBase(BaseDataType.REAL);
+    public static final HMTypeBase REAL = new HMTypeBase(BaseDataType.DOUBLE);
     public static final HMTypeBase BOOL = new HMTypeBase(BaseDataType.BOOL);
     public static final HMTypeBase STR = new HMTypeBase(BaseDataType.STR);
     public static final HMTypeBase VOID = new HMTypeBase(BaseDataType.VOID);
@@ -29,7 +29,7 @@ public class HMTypeBase extends HMType {
         switch (this.baseType) {
             case INT:
                 return "Integer";
-            case REAL:
+            case DOUBLE:
                 return "Double";
             case BOOL:
                 return "Boolean";
